Update functions and fix errors
